requirements.txt
discord.py==1.6.0
pytube==10.6.0